third_party/WebKit/LayoutTests/inspector/sources/debugger/script-formatter-breakpoints-3.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script src="resources/unformatted3.js"></script>
<script>
var test = function()
{
    WebInspector.breakpointManager._storage._breakpoints = {};
    var panel = WebInspector.inspectorView.showPanel("sources");
    var scriptFormatter;

    InspectorTest.runDebuggerTestSuite([
        function testSetup(next)
        {
            InspectorTest.scriptFormatter().then(function(sf) {
                scriptFormatter = sf;
                next();
            });
        },

        function testBreakpointsSetInFormattedAndRemoveInOriginalSource(next)
        {
            InspectorTest.showScriptSource("unformatted3.js", didShowScriptSource);

            function didShowScriptSource(frame)
            {
                InspectorTest.addSniffer(WebInspector.ScriptFormatterEditorAction.prototype, "_updateButton", uiSourceCodeScriptFormatted);
                scriptFormatter._toggleFormatScriptSource();
            }
                
            function uiSourceCodeScriptFormatted()
            {
                var formattedSourceFrame = panel.visibleView;
                InspectorTest.setBreakpoint(formattedSourceFrame, 3, "", true);
                InspectorTest.waitUntilPaused(pausedInF2);
                InspectorTest.evaluateInPageWithTimeout("f2()");
            }

            function pausedInF2(callFrames)
            {
                InspectorTest.dumpBreakpointSidebarPane("while paused in pretty printed");
                scriptFormatter._discardFormattedUISourceCodeScript(panel.visibleView.uiSourceCode());
                InspectorTest.dumpBreakpointSidebarPane("while paused in raw");
                // No need to remove breakpoint since formattedUISourceCode was removed.
                InspectorTest.resumeExecution(next);
            }
        }
    ]);

}
</script>
</head>
<body onload="runTest()">
<p>Tests the script formatting is working fine with breakpoints.
</p>
</body>
</html>
