third_party/WebKit/LayoutTests/inspector/sources/debugger/script-formatter-breakpoints-3-expected.txt

Tests the script formatting is working fine with breakpoints.


Running: testBreakpointsSetInFormattedAndRemoveInOriginalSource
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed

unformatted3.js:formatted:4    var b = 1;
Breakpoint sidebar pane while paused in raw

Script execution resumed.
